Assembler quase pronto :)

+++ src/main/java/com/vm/ps/vmproject_ps/Registers.java
@@ -23,7 +23,7 @@ public class Registers {
     }
     public void clearAll(){
         for(String i: registerSet.keySet()){
-            registerSet.get(i).setValue(0);
+            registerSet.get(i).setValue((short) 0);
         }
     }
 }
